Fix relationship loading bug when using STI.
When using STI, the entity map would previously fail to detect a relationship
definition because it was only checking that the name of the first param (in the definition)
was of the exact same class as the entity for the map. However, when using STI this won't always be
the case and so we need to also allow for the entity to be a subclass of the param.